+++ django_telegrambot/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '0.2.7'
+__version__ = '1.0.0'
 default_app_config = 'django_telegrambot.apps.DjangoTelegramBot'