+++ django_telegrambot/views.py
@@ -20,7 +20,7 @@ logger = logging.getLogger(__name__)
 @staff_member_required
 def home(request):
     bot_list = DjangoTelegramBot.bots
-    context = {'bot_list': bot_list, 'update_mode':settings.TELEGRAM_BOT_MODE}
+    context = {'bot_list': bot_list, 'update_mode':settings.DJANGO_TELEGRAMBOT.get('MODE', 'WEBHOOK')}
     return render(request, 'django_telegrambot/index.html', context)