+++ sampleproject/bot/views.py
@@ -1,8 +1,9 @@
 from django.shortcuts import render
+from django.conf import settings
 from django_telegrambot.apps import DjangoTelegramBot
 
 # Create your views here.
 def index(request):
     bot_list = DjangoTelegramBot.bots
-    context = {'bot_list': bot_list}
+    context = {'bot_list': bot_list, 'update_mode':settings.TELEGRAM_BOT_MODE}
     return render(request, 'bot/index.html', context)